move PublicRequests to own file, update imports
