+++ src/lib/RequestTypes.ts
@@ -1,6 +1,14 @@
-import { WalletType, WalletInfoEntry } from './WalletInfo';
 import { TX_VALIDITY_WINDOW, TX_MIN_VALIDITY_DURATION } from './Constants';
 import { State } from '@nimiq/rpc';
+import {
+    BasicRequest,
+    SimpleRequest,
+    SignTransactionRequest,
+    CheckoutRequest,
+    SignMessageRequest,
+    RenameRequest,
+    RpcRequest,
+} from './PublicRequestTypes';
 
 export enum RequestType {
     LIST = 'list',
@@ -19,36 +27,13 @@ export enum RequestType {
     CHOOSE_ADDRESS = 'choose-address',
 }
 
-export interface BasicRequest {
-    appName: string;
-}
-
 export interface ParsedBasicRequest {
     kind: RequestType;
     appName: string;
 }
 
-export interface SimpleRequest extends BasicRequest {
-    accountId: string;
-}
-
 export interface ParsedSimpleRequest extends ParsedBasicRequest {
     walletId: string;
-}
-
-export interface SimpleResult {
-    success: true;
-}
-
-export interface SignTransactionRequest extends SimpleRequest {
-    sender: string;
-    recipient: string;
-    recipientType?: Nimiq.Account.Type;
-    value: number;
-    fee?: number;
-    extraData?: Uint8Array;
-    flags?: number;
-    validityStartHeight: number; // FIXME To be made optional when accounts manager has its own network
 }
 
 export interface ParsedSignTransactionRequest extends ParsedSimpleRequest {
@@ -63,38 +48,6 @@ export interface ParsedSignTransactionRequest extends ParsedSimpleRequest {
     validityStartHeight: number; // FIXME To be made optional when accounts manager has its own network
 }
 
-export interface SignedTransaction {
-    serializedTx: string; // HEX
-    hash: string; // HEX
-
-    raw: {
-        signerPublicKey: Uint8Array;
-        signature: Uint8Array;
-
-        sender: string; // Userfriendly address
-        senderType: Nimiq.Account.Type;
-        recipient: string; // Userfriendly address
-        recipientType: Nimiq.Account.Type;
-        value: number; // Luna
-        fee: number; // Luna
-        validityStartHeight: number;
-        extraData: Uint8Array;
-        flags: number;
-        networkId: number;
-    };
-}
-
-export interface CheckoutRequest extends BasicRequest {
-    shopLogoUrl?: string;
-    recipient: string;
-    recipientType?: Nimiq.Account.Type;
-    value: number;
-    fee?: number;
-    extraData?: Uint8Array;
-    flags?: number;
-    validityDuration?: number;
-}
-
 export interface ParsedCheckoutRequest extends ParsedBasicRequest {
     shopLogoUrl?: string;
     recipient: Nimiq.Address;
@@ -106,68 +59,23 @@ export interface ParsedCheckoutRequest extends ParsedBasicRequest {
     validityDuration: number;
 }
 
-export interface SignMessageRequest extends BasicRequest {
-    accountId?: string;
-    signer?: string;
-    message: string | Uint8Array;
-}
-
 export interface ParsedSignMessageRequest extends ParsedBasicRequest {
     walletId?: string;
     signer?: Nimiq.Address;
     message: string | Uint8Array;
 }
 
-export interface SignedMessage {
-    signer: string; // Userfriendly address
-    signerPubKey: Uint8Array;
-    signature: Uint8Array;
-    message: string | Uint8Array;
-}
-
-export interface Address {
-    address: string; // Userfriendly address
-    label: string;
-}
-
-export interface Account {
-    accountId: string;
-    label: string;
-    type: WalletType;
-    addresses: Address[];
-}
-
-export interface RenameRequest extends SimpleRequest {
-    address?: string;
-}
-
 export interface ParsedRenameRequest extends ParsedSimpleRequest {
     address?: string;
 }
 
-export type ListResult = WalletInfoEntry[];
-
 // Discriminated Unions
-export type RpcRequest = SignTransactionRequest
-                       | CheckoutRequest
-                       | BasicRequest
-                       | SimpleRequest
-                       | RenameRequest
-                       | SignMessageRequest;
-
 export type ParsedRpcRequest = ParsedSignTransactionRequest
                              | ParsedCheckoutRequest
                              | ParsedBasicRequest
                              | ParsedSimpleRequest
                              | ParsedRenameRequest
                              | ParsedSignMessageRequest;
-
-export type RpcResult = SignedTransaction
-                      | Account
-                      | SimpleResult
-                      | Address
-                      | SignedMessage
-                      | ListResult;
 
 export class AccountsRequest {
     public static parse(request: RpcRequest, state: State, requestType?: RequestType): ParsedRpcRequest | null {
