+++ src/views/Export.vue
@@ -2,14 +2,14 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import { ParsedExportRequest } from '../lib/RequestTypes';
+import { ParsedSimpleRequest } from '../lib/RequestTypes';
 import { SimpleRequest } from '@nimiq/keyguard-client';
 import { WalletStore } from '@/lib/WalletStore';
 import { Static } from '../lib/StaticStore';
 
 @Component
 export default class Export extends Vue {
-    @Static private request!: ParsedExportRequest;
+    @Static private request!: ParsedSimpleRequest;
 
     public async created() {
         const wallet = await WalletStore.Instance.get(this.request.walletId);
