+++ src/views/SignMessageSuccess.vue
@@ -14,7 +14,8 @@
 <script lang="ts">
 import { Component, Emit, Vue } from 'vue-property-decorator';
 import { State } from 'vuex-class';
-import { ParsedSignMessageRequest, SignedMessage } from '../lib/RequestTypes';
+import { ParsedSignMessageRequest } from '../lib/RequestTypes';
+import { SignedMessage } from '../lib/PublicRequestTypes';
 import KeyguardClient from '@nimiq/keyguard-client';
 import { Static } from '@/lib/StaticStore';
 import Loader from '../components/Loader.vue';
