+++ src/views/SignTransactionSuccess.vue
@@ -16,7 +16,7 @@
 import { Component, Vue } from 'vue-property-decorator';
 import Network from '@/components/Network.vue';
 import { SmallPage } from '@nimiq/vue-components';
-import { SignedTransaction } from '../lib/RequestTypes';
+import { SignedTransaction } from '../lib/PublicRequestTypes';
 import { State } from 'vuex-class';
 import { Static } from '../lib/StaticStore';
 import Loader from '../components/Loader.vue';
