+++ src/views/Signup.vue
@@ -2,14 +2,14 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import { ParsedSignupRequest } from '../lib/RequestTypes';
+import { ParsedBasicRequest } from '../lib/RequestTypes';
 import { Static } from '../lib/StaticStore';
 import { DEFAULT_KEY_PATH } from '@/lib/Constants';
 import KeyguardClient from '@nimiq/keyguard-client';
 
 @Component
 export default class Signup extends Vue {
-    @Static private request!: ParsedSignupRequest;
+    @Static private request!: ParsedBasicRequest;
 
     public created() {
         const request: KeyguardClient.CreateRequest = {
